moved teloxide::utils::command::tests to tests/command.rs
